src/third_party/blink/web_tests/fast/forms/file/file-input-webkitdirectory-key-space-expected.txt

CONSOLE MESSAGE: line 5: FileChooser: opened; mode=kUploadFolder
Ensure pressing Space key with focus on an INPUT element with TYPE=FILE and WEBKITDIRECTORY launches a file chooser. Automated test passes if 'FileChooser: opened' was logged.
